Fix task scheduler test.

+++ sop_src/utils/byteq.c
@@ -18,6 +18,7 @@
 #include "utils/helper_macros.h"
 #include "utils/platform_log.h"
 
+
 #define ASSERT LOG_ASSERT
 #define ASSERT_FN LOG_ASSERT_FN
 
@@ -527,9 +528,8 @@ void ByteQSetRdIdxFromPointer(ByteQ_t *const pQ, void *pRdPtr) {
   }
   {
     const bq_t *const pRd8 = (const bq_t *)pRdPtr;
-    //(Lint):Note 946: Relational or subtract operator applied to pointers
-    //[MISRA 2004 Rule 17.3]
-    ssize_t newRdIdx = pRd8 - pQ->pfBuf; // lint !e946
+
+    int newRdIdx = pRd8 - pQ->pfBuf; // lint !e946
 
     // Check for within range.
     if ((newRdIdx >= 0) && (newRdIdx <= (int)pQ->nBufSz)) // lint !e574 !e737
